Moved chat-management services to "chat" package
