+++ messaging-service/internal/application/services/chat/group_photo_service.go
@@ -1,4 +1,4 @@
-package services
+package chat
 
 import (
 	"context"
