+++ messaging-service/internal/application/services/chat/personal_chat_service_test.go
@@ -1,4 +1,4 @@
-package services
+package chat
 
 import (
 	"context"
